backend/pyproject.toml
[project]
name = "propertyfish-backend"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "psycopg2-binary",
  "sqlalchemy",
  "python-dotenv",
  "geojson",
]

[tool.pytest.ini_options]
pythonpath = ["backend/app"]
